+++ Changes.md
@@ -7,6 +7,7 @@ Change Log in version 3.10.1.4 (2021022304)
 5. Settings in the CSS can now be pre-processed before passed to the SCSS compiler.  Thus enabling further transition to SCSS of the CSS.
    This will reduce the amount of source 'CSS' and help to spot duplication and mistakes.
 6. Added extra information in the settings tabs about maturity of the release.
+7. Fix 'Userdata cache not used' - #247.
 
 Change Log in version 3.10.1.3 (2021022303)
 ===========================================